requirements.txt
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.15.0
pillow==10.2.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2024.1
